UI changes to better reflect what workouts can be modified
added a finish button to end workouts, both in the `/workout` route and `/user` recent workout list
in `/user` you can now click old workouts and view them but not edit. removed the continue button in favor of selecting the workout to expand it with the current open workout being the only editable one.
added the ability to delete open workouts with nothing added.